+++ CHANGELOG.md
@@ -11,11 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added additional variants to `shader_database.Operation`.
 * Added `shader_database.Dependency.int` variant.
 * Added field `unk` to `LodData` to better preserve in game data.
+* Added additional variants to `material.RenderPassType`.
+* Added field `alt_textures` to `material.Material`.
 
 ### Changed
 * Changed numpy arrays to be row-major to properly match conventions.
 * Moved `bone_names` field for `skinning.SkinWeights` to parameters for `to_influences` and `add_influences` methods.
 * Renamed `shader_database.Dependency.constant` to `shader_database.Dependency.float`.
+* Renamed field `material.Material.m_unks2_2` to `m_unks2`.
 
 ### Removed
 * Removed field `unk5` from `LodItem`.